+++ packages/store/src/index.ts
@@ -1,3 +1,4 @@
+export * from './abstract-collection';
 export * from './collection';
 export * from './store';
 export * from './store-service';