+++ src/docs/.vuepress/config.js
@@ -51,8 +51,8 @@ function getGuideSidebar() {
 }
 
 module.exports = {
-  title: 'VueenT',
-  description: 'A library which may help to create a scalable Vue.js application',
+  title: 'VueEnt',
+  description: ' A library for building scalable Vue.js applications',
   base: '/vueent/',
   head: [
     ['link', { rel: 'icon', href: '/img/logo.png' }]